[MOB-10001] Remove Media Projection from Android (#281)
When invoking Instabug on Android, a "scary" dialog appears, asking for
permission to record screen. This was due to the way we captured
screenshots on Android, using Media Projection, since other
alternatives didn't work with Flutter.

This PR aims at providing an alternative approach for Media Projection,
using FlutterRenderer to capture screenshots. This implicitly fixes
Repro Steps on Android as well.