fix: skip is APM enabled check when starting a ui trace
On Android, `FlagsConfig.apm.isEnabled` isn't implemented correctly so we skip the isApmEnabled check on Android and only check on iOS when calling `startUiTrace`.
This is a temporary fix until we implement the isEnabled check correctly. We need to fix this in the future and implement APM.isEnabled correctly to avoid discrepancy.